requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
autopep8==1.5.6
boto3==1.17.87
botocore==1.20.87
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.4
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extensions==3.1.2
django-health-check==3.16.4
django-sendgrid-v5==0.9.1
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-serializer-extensions==2.0.1
future==0.18.2
hashids==1.3.1
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
sendgrid==6.7.0
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.2
urllib3==1.26.4
